Split step service in two
